chore(tests): harmonize naming and stubs
