+++ tests/unit/async_scan_test.rb
@@ -4,7 +4,8 @@ require 'minitest/autorun'
 require 'tmpdir'
 require 'ostruct'
 
-$LOADED_FEATURES << 'sketchup.rb'
+$LOAD_PATH.unshift File.expand_path('../../test/stubs', __dir__)
+require 'sketchup'
 
 # --- Stubs for SketchUp API ---
 module UI
@@ -65,50 +66,6 @@ end
 
 module Geom
   Z_AXIS = [0, 0, 1].freeze
-end
-
-module Sketchup
-  class ComponentInstance; end
-  class Group < ComponentInstance; end
-  class ModelObserver; end
-  class SelectionObserver; end
-
-  class Layer
-    def visible? = true
-    def name = ''
-  end
-
-  class Entities
-    def initialize(list)
-      @list = list
-    end
-
-    def to_a
-      @list
-    end
-  end
-
-  class Model
-    attr_reader :entities, :selection, :layers
-
-    def initialize(entities)
-      @entities = Entities.new(entities)
-      @selection = Entities.new([])
-      @layers = []
-    end
-  end
-
-  def self.active_model
-    @model
-  end
-
-  def self.active_model=(model)
-    @model = model
-  end
-
-  def self.temp_dir
-    Dir.tmpdir
-  end
 end
 
 class MockDefinition
