+++ tests/unit/detach_observers_test.rb
