+++ tests/unit/scanner_test.rb
